Updated Vulkan Module Organization
