--- Source/Rasterer/Graphics API/RasterContext.cpp
@@ -1,7 +1,0 @@
-#include "RasterContext.h"
-
-RasterContext::RasterContext(RasterAPI rasterAPI = RasterAPI::NO_API) :
-	rasterAPI_(rasterAPI)
-{
-}
-
