--- Source/Rasterer/Graphics API/RenderPass.cpp
@@ -1,1 +1,0 @@
-#include "RenderPass.h"
