--- Source/Rasterer/Graphics API/RenderPass.h
@@ -1,17 +1,0 @@
-#pragma once
-
-class RenderPass {
-
-public:
-
-	RenderPass() = default;
-	virtual ~RenderPass() = default;
-
-	RenderPass(const RenderPass&) = delete;
-	RenderPass(RenderPass&& o) noexcept = delete;
-
-	RenderPass& operator=(const RenderPass&) = delete;
-	RenderPass& operator=(RenderPass&& other) noexcept = delete;
-
-
-};
