--- Source/Rasterer/Graphics API/Shader.cpp
@@ -1,6 +1,0 @@
-#include "Shader.h"
-
-Shader::Shader(const std::string& fileName):
-fileName_(fileName)
-{
-}
