--- Source/Rasterer/Graphics API/Shader/FragmentShader.cpp
@@ -1,7 +1,0 @@
-#include "FragmentShader.h"
-
-FragmentShader::FragmentShader(const std::string& fileName) :
-	Shader(fileName)
-{
-	Shader::shaderType_ = ShaderType::Fragment;
-}
