--- Source/Rasterer/Graphics API/Shader/VertexShader.cpp
@@ -1,7 +1,0 @@
-#include "VertexShader.h"
-
-VertexShader::VertexShader(const std::string& fileName) :
-	Shader(fileName)
-{
-	Shader::shaderType_ = ShaderType::Vertex;
-}
