--- Source/Rasterer/Graphics API/Surface.cpp
@@ -1,1 +1,0 @@
-#include "Surface.h"
