--- Source/Rasterer/Graphics API/SwapChain.cpp
@@ -1,1 +1,0 @@
-#include "SwapChain.h"
