+++ Source/Rasterer/Modules/Vulkan/Shader/VulkanFragmentShader.cpp
@@ -1,8 +1,6 @@
 #include "VulkanFragmentShader.h"
 
 VulkanFragmentShader::VulkanFragmentShader(EntityManager& entityManger, Entity device, const std::string& fileName) :
-	Shader(fileName),
-	FragmentShader(fileName),
 	VulkanShader(entityManger, device,fileName)
 {
 }
