+++ Source/Rasterer/Modules/Vulkan/VulkanRasterBackend.h
@@ -2,11 +2,14 @@
 
 #include "Rasterer/RasterBackend.h"
 
-class VulkanRasterBackend final: public RasterBackend {
+#include <vulkan/vulkan.hpp>
+
+
+class VulkanRasterBackend final : public RasterBackend {
 
 public:
 
-	VulkanRasterBackend() = default;
+	VulkanRasterBackend();
 	~VulkanRasterBackend() override = default;
 
 	VulkanRasterBackend(const VulkanRasterBackend &) = delete;
@@ -19,5 +22,26 @@ public:
 	void DrawIndirect() override;
 
 	std::shared_ptr<RasterDevice> CreateDevice() override;
+
+	vk::Instance& GetInstance();
+
+private:
+
+
+	std::vector<char const*> requiredDeviceExtensions_;
+	std::vector<char const*> requiredInstanceExtensions_;
+	std::vector<const char*> validationLayers;
+
+	vk::Instance instance_;
+	std::vector<vk::PhysicalDevice> physicalDevices_;
+
+	//Debug related state 
+	//TODO: Should be conditionally included when the class is only debug mode.
+
+	static VkBool32 DebugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT, VkDebugUtilsMessageTypeFlagsEXT, const VkDebugUtilsMessengerCallbackDataEXT*, void*);
+
+	VkDebugUtilsMessengerEXT debugMessenger_;
+	vk::DispatchLoaderDynamic dispatcher_;
+	vk::DebugReportCallbackEXT callback_;
 
 };
